# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "CancelTripRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "CancelTripRequest",
        "type": "object"
      },
...
      "Coordinate": {
        "properties": {
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng",
            "type": "number"
          }
        },
        "required": [
          "lat",
          "lng"
        ],
        "title": "Coordinate",
        "type": "object"
      },
...
      "FareBreakdown": {
        "description": "Breakdown đầy đủ theo SPEC 4.4.\n\nLưu ý phân bổ: `insurance_fee` và `payment_gateway_fee` được trừ TRONG phần\n`platform_commission` (nền tảng giữ), không thu thêm của rider.",
        "properties": {
          "base_fee": {
            "title": "Base Fee",
            "type": "string"
          },
          "distance_fee": {
            "title": "Distance Fee",
            "type": "string"
          },
          "driver_payout": {
            "title": "Driver Payout",
            "type": "string"
          },
          "final_fare": {
            "title": "Final Fare",
            "type": "string"
          },
          "insurance_fee": {
            "title": "Insurance Fee",
            "type": "string"
          },
          "payment_gateway_fee": {
            "title": "Payment Gateway Fee",
            "type": "string"
          },
          "pickup_surcharge": {
            "title": "Pickup Surcharge",
            "type": "string"
          },
          "platform_commission": {
            "title": "Platform Commission",
            "type": "string"
          },
          "subtotal": {
            "title": "Subtotal",
            "type": "string"
          },
          "time_fee": {
            "title": "Time Fee",
            "type": "string"
          }
        },
        "required": [
          "base_fee",
          "distance_fee",
          "time_fee",
          "pickup_surcharge",
          "subtotal",
          "final_fare",
          "driver_payout",
          "platform_commission",
          "insurance_fee",
          "payment_gateway_fee"
        ],
        "title": "FareBreakdown",
        "type": "object"
      },
...
      "FraudDetectedBy": {
        "enum": [
          "system",
          "report"
        ],
        "title": "FraudDetectedBy",
        "type": "string"
      },
      "FraudIncidentOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "details": {
            "title": "Details",
            "type": "object"
          },
          "detected_by": {
            "$ref": "#/components/schemas/FraudDetectedBy"
          },
          "driver_id": {
            "format": "uuid",
            "title": "Driver Id",
            "type": "string"
          },
          "fraud_type": {
            "$ref": "#/components/schemas/FraudType"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "penalty_amount": {
            "title": "Penalty Amount",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/FraudSeverity"
          },
          "trip_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Id"
          }
        },
        "required": [
          "id",
          "trip_id",
          "driver_id",
          "fraud_type",
          "detected_by",
          "severity",
          "penalty_amount",
          "details",
          "created_at"
        ],
        "title": "FraudIncidentOut",
        "type": "object"
      },
      "FraudReviewItemOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "details": {
            "title": "Details",
            "type": "object"
          },
          "driver_id": {
            "format": "uuid",
            "title": "Driver Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "signal_score": {
            "title": "Signal Score",
            "type": "number"
          },
          "status": {
            "$ref": "#/components/schemas/FraudReviewStatus"
          }
        },
        "required": [
          "id",
          "driver_id",
          "reason",
          "signal_score",
          "details",
          "status",
          "created_at"
        ],
        "title": "FraudReviewItemOut",
        "type": "object"
      },
      "FraudReviewStatus": {
        "enum": [
          "pending",
          "cleared",
          "confirmed"
        ],
        "title": "FraudReviewStatus",
        "type": "string"
      },
      "FraudSeverity": {
        "enum": [
          "warning",
          "account_locked"
        ],
        "title": "FraudSeverity",
        "type": "string"
      },
      "FraudType": {
        "enum": [
          "ghost_trip",
          "route_deviation",
          "off_app_payment",
          "driver_swap"
        ],
        "title": "FraudType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewDecision": {
        "properties": {
          "confirmed": {
            "title": "Confirmed",
            "type": "boolean"
          },
          "trip_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Id"
          }
        },
        "required": [
          "confirmed"
        ],
        "title": "ReviewDecision",
        "type": "object"
      },
...
      "TimeBand": {
        "enum": [
          "normal",
          "night",
          "peak"
        ],
        "title": "TimeBand",
        "type": "string"
      },
...
      "TripCreate": {
        "properties": {
          "dropoff": {
            "$ref": "#/components/schemas/Coordinate"
          },
          "dropoff_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dropoff Address"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "pickup": {
            "$ref": "#/components/schemas/Coordinate"
          },
          "pickup_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pickup Address"
          },
          "restaurant_partner_qr_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restaurant Partner Qr Token"
          }
        },
        "required": [
          "pickup",
          "dropoff"
        ],
        "title": "TripCreate",
        "type": "object"
      },
      "TripCreateResponse": {
        "properties": {
          "estimate": {
            "$ref": "#/components/schemas/FareBreakdown"
          },
          "trip": {
            "$ref": "#/components/schemas/TripOut"
          }
        },
        "required": [
          "trip",
          "estimate"
        ],
        "title": "TripCreateResponse",
        "type": "object"
      },
      "TripOut": {
        "properties": {
          "cancellation_fee": {
            "title": "Cancellation Fee",
            "type": "string"
          },
          "cancellation_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancellation Reason"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "distance_km": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Km"
          },
          "driver_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Id"
          },
          "driver_payout": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Payout"
          },
          "dropoff_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dropoff Address"
          },
          "dropoff_lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dropoff Lat"
          },
          "dropoff_lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dropoff Lng"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "estimated_fare": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Fare"
          },
          "final_fare": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Fare"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "insurance_fee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Insurance Fee"
          },
          "insurance_voided": {
            "title": "Insurance Voided",
            "type": "boolean"
          },
          "matched_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched At"
          },
          "optimal_distance_km": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optimal Distance Km"
          },
          "pickup_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pickup Address"
          },
          "pickup_lat": {
            "title": "Pickup Lat",
            "type": "number"
          },
          "pickup_lng": {
            "title": "Pickup Lng",
            "type": "number"
          },
          "pickup_surcharge": {
            "title": "Pickup Surcharge",
            "type": "string"
          },
          "platform_commission": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform Commission"
          },
          "qr_verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qr Verified At"
          },
          "requested_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested At"
          },
          "rider_id": {
            "format": "uuid",
            "title": "Rider Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/TripStatus"
          },
          "time_band": {
            "$ref": "#/components/schemas/TimeBand"
          }
        },
        "required": [
          "id",
          "rider_id",
          "driver_id",
          "status",
          "time_band",
          "pickup_lat",
          "pickup_lng",
          "pickup_address",
          "dropoff_lat",
          "dropoff_lng",
          "dropoff_address",
          "estimated_fare",
          "final_fare",
          "distance_km",
          "optimal_distance_km",
          "duration_minutes",
          "pickup_surcharge",
          "platform_commission",
          "driver_payout",
          "insurance_fee",
          "insurance_voided",
          "cancellation_fee",
          "qr_verified_at",
          "requested_at",
          "matched_at",
          "started_at",
          "completed_at",
          "cancelled_at",
          "cancellation_reason"
        ],
        "title": "TripOut",
        "type": "object"
      },
      "TripStatus": {
        "enum": [
          "requested",
          "matching",
          "matched",
          "driver_arriving",
          "qr_verified",
          "in_progress",
          "completed",
          "cancelled_by_rider",
          "cancelled_by_driver",
          "no_driver_found"
        ],
        "title": "TripStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Dán access_token từ /api/v1/auth/verify-otp",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "GoAn Backend",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/fraud/incidents": {
      "get": {
        "operationId": "list_incidents_api_v1_admin_fraud_incidents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FraudIncidentOut"
                  },
                  "title": "Response List Incidents Api V1 Admin Fraud Incidents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Incidents",
        "tags": [
          "fraud"
        ]
      }
    },
    "/api/v1/admin/fraud/review-queue": {
      "get": {
        "description": "Tín hiệu nghi ngờ thanh toán ngoài app, chờ admin xử lý thủ công (SPEC 7.3).",
        "operationId": "list_review_queue_api_v1_admin_fraud_review_queue_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FraudReviewItemOut"
                  },
                  "title": "Response List Review Queue Api V1 Admin Fraud Review Queue Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Review Queue",
        "tags": [
          "fraud"
        ]
      }
    },
    "/api/v1/admin/fraud/review-queue/{item_id}/decide": {
      "post": {
        "operationId": "decide_review_item_api_v1_admin_fraud_review_queue__item_id__decide_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FraudReviewItemOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Decide Review Item",
        "tags": [
          "fraud"
        ]
      }
    },
...
    "/api/v1/trips": {
      "post": {
        "operationId": "create_trip_api_v1_trips_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/api/v1/trips/{trip_id}": {
      "get": {
        "operationId": "get_trip_api_v1_trips__trip_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/api/v1/trips/{trip_id}/cancel": {
      "post": {
        "operationId": "cancel_trip_api_v1_trips__trip_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelTripRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Trip",
        "tags": [
          "trips"
        ]
      }
    },
...
  }
}